requirements.txt
Flask==2.0.2
Werkzeug==2.0.2
pdfminer.six==20201018
python-docx==0.8.11
nltk==3.6.3
scikit-learn==0.24.2
spacy
requests